backend/requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.0
pymysql>=1.1.0
alembic>=1.12.0
redis>=5.0.0

# Web Scraping
scrapy>=2.11.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0
fake-useragent>=1.4.0

# NLP & ML
jieba>=0.42.1
scikit-learn>=1.3.0
sentence-transformers>=2.2.0

# LLM & API
openai>=1.3.0
httpx>=0.25.0
anthropic>=0.7.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-multipart>=0.0.6

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
